se reorganizo la estructura de las pantallas de la sala
