+++ src/components/room/lobby/views/LobbyView.vue
@@ -80,13 +80,13 @@ export default {
 <script setup>
 
 import { ref, onMounted, provide, watch, computed } from 'vue';
-import { useWebSocketStore } from '../stores/socketStore'
+import { useWebSocketStore } from '../../stores/socketStore'
 import { storeToRefs } from 'pinia'
 import { useRouter } from 'vue-router'
 import ColorThief from 'colorthief';
-import TabChat from './components/TabChat.vue';
-import TabInfo from './components/TabInfo.vue';
-import TabUsers from './components/TabUsers.vue';
+import TabChat from '../components/TabChat.vue';
+import TabInfo from '../components/TabInfo.vue';
+import TabUsers from '../components/TabUsers.vue';
 import { Icon } from "@iconify/vue";
 
 const socketStore = useWebSocketStore();
